Reached stage : Finalizing build. Testing and minor changes.

+++ app/src/main/java/lm/pkp/com/landmap/permission/PermissionsDBHelper.java
@@ -23,7 +23,7 @@ import lm.pkp.com.landmap.user.UserElement;
 public class PermissionsDBHelper extends SQLiteOpenHelper {
 
     public static final String DATABASE_NAME = "landmap.db";
-    private AsyncTaskCallback callback = null;
+    private AsyncTaskCallback callback;
 
     public static final String ACCESS_TABLE_NAME = "area_access";
     public static final String ACCESS_COLUMN_AREA_ID = "area_id";
@@ -31,52 +31,52 @@ public class PermissionsDBHelper extends SQLiteOpenHelper {
     public static final String ACCESS_COLUMN_FUNCTION_CODE = "function_code";
 
     public PermissionsDBHelper(Context context, AsyncTaskCallback callback) {
-        super(context, DATABASE_NAME, null, 1);
+        super(context, PermissionsDBHelper.DATABASE_NAME, null, 1);
         this.callback = callback;
     }
 
     public PermissionsDBHelper(Context context) {
-        super(context, DATABASE_NAME, null, 1);
+        super(context, PermissionsDBHelper.DATABASE_NAME, null, 1);
     }
 
     @Override
     public void onCreate(SQLiteDatabase db) {
         db.execSQL(
-                "create table " + ACCESS_TABLE_NAME + "(" +
-                        ACCESS_COLUMN_AREA_ID + " text," +
-                        ACCESS_COLUMN_USER_ID + " text, " +
-                        ACCESS_COLUMN_FUNCTION_CODE + " text)"
+                "create table " + PermissionsDBHelper.ACCESS_TABLE_NAME + "(" +
+                        PermissionsDBHelper.ACCESS_COLUMN_AREA_ID + " text," +
+                        PermissionsDBHelper.ACCESS_COLUMN_USER_ID + " text, " +
+                        PermissionsDBHelper.ACCESS_COLUMN_FUNCTION_CODE + " text)"
         );
     }
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("DROP TABLE IF EXISTS " + ACCESS_TABLE_NAME);
-        onCreate(db);
+        db.execSQL("DROP TABLE IF EXISTS " + PermissionsDBHelper.ACCESS_TABLE_NAME);
+        this.onCreate(db);
     }
 
     public PermissionElement insertPermission(PermissionElement pe) {
-        SQLiteDatabase db = this.getWritableDatabase();
+        SQLiteDatabase db = getWritableDatabase();
 
         ContentValues contentValues = new ContentValues();
-        contentValues.put(ACCESS_COLUMN_AREA_ID, pe.getAreaId());
-        contentValues.put(ACCESS_COLUMN_FUNCTION_CODE, pe.getFunctionCode());
-        contentValues.put(ACCESS_COLUMN_USER_ID, pe.getUserId());
+        contentValues.put(PermissionsDBHelper.ACCESS_COLUMN_AREA_ID, pe.getAreaId());
+        contentValues.put(PermissionsDBHelper.ACCESS_COLUMN_FUNCTION_CODE, pe.getFunctionCode());
+        contentValues.put(PermissionsDBHelper.ACCESS_COLUMN_USER_ID, pe.getUserId());
 
-        db.insert(ACCESS_TABLE_NAME, null, contentValues);
+        db.insert(PermissionsDBHelper.ACCESS_TABLE_NAME, null, contentValues);
         db.close();
         return pe;
     }
 
     public void insertPermissionsToServer(String targetUser, String functionCodes) {
-        LMSRestAsyncTask task = new LMSRestAsyncTask(callback);
-        task.execute(preparePostParams("insert", targetUser, functionCodes));
+        LMSRestAsyncTask task = new LMSRestAsyncTask(this.callback);
+        task.execute(this.preparePostParams("insert", targetUser, functionCodes));
     }
 
     private JSONObject preparePostParams(String queryType, String targetUser, String functionCodes) {
         JSONObject postParams = new JSONObject();
-        final AreaElement areaElement = AreaContext.INSTANCE.getAreaElement();
-        final UserElement userElement = UserContext.getInstance().getUserElement();
+        AreaElement areaElement = AreaContext.INSTANCE.getAreaElement();
+        UserElement userElement = UserContext.getInstance().getUserElement();
         try {
             postParams.put("requestType", "AreaShare");
             postParams.put("query_type", queryType);
@@ -91,24 +91,24 @@ public class PermissionsDBHelper extends SQLiteOpenHelper {
     }
 
     public void deletePermissionsByAreaId(String areaId) {
-        SQLiteDatabase db = this.getWritableDatabase();
-        db.execSQL("DELETE FROM " + ACCESS_TABLE_NAME + " WHERE "
-                + ACCESS_COLUMN_AREA_ID + " = '" + areaId + "'");
+        SQLiteDatabase db = getWritableDatabase();
+        db.execSQL("DELETE FROM " + PermissionsDBHelper.ACCESS_TABLE_NAME + " WHERE "
+                + PermissionsDBHelper.ACCESS_COLUMN_AREA_ID + " = '" + areaId + "'");
         db.close();
     }
 
     public Map<String, PermissionElement> fetchPermissionsByAreaId(String areaId) {
-        SQLiteDatabase db = this.getWritableDatabase();
+        SQLiteDatabase db = getWritableDatabase();
         Map<String, PermissionElement> perMap = new HashMap<>();
-        Cursor cursor = db.rawQuery("select * from " + ACCESS_TABLE_NAME
-                + " WHERE " + ACCESS_COLUMN_AREA_ID + "=?", new String[]{areaId});
+        Cursor cursor = db.rawQuery("select * from " + PermissionsDBHelper.ACCESS_TABLE_NAME
+                + " WHERE " + PermissionsDBHelper.ACCESS_COLUMN_AREA_ID + "=?", new String[]{areaId});
         if (cursor != null) {
             cursor.moveToFirst();
             while (cursor.isAfterLast() == false) {
                 PermissionElement pe = new PermissionElement();
-                pe.setUserId(cursor.getString(cursor.getColumnIndex(ACCESS_COLUMN_USER_ID)));
+                pe.setUserId(cursor.getString(cursor.getColumnIndex(PermissionsDBHelper.ACCESS_COLUMN_USER_ID)));
                 pe.setAreaId(areaId);
-                final String functionCode = cursor.getString(cursor.getColumnIndex(ACCESS_COLUMN_FUNCTION_CODE));
+                String functionCode = cursor.getString(cursor.getColumnIndex(PermissionsDBHelper.ACCESS_COLUMN_FUNCTION_CODE));
                 pe.setFunctionCode(functionCode);
                 perMap.put(functionCode, pe);
                 cursor.moveToNext();
@@ -124,7 +124,7 @@ public class PermissionsDBHelper extends SQLiteOpenHelper {
     }
 
     public void finalizeTaskCompletion() {
-        callback.taskCompleted("");
+        this.callback.taskCompleted("");
     }
 
 }
